refactor: extract helpers/logs.py (Phase 6.2)
Move the four filesystem log helpers out of dashboard.py's module-level
bag into helpers/logs.py:

  _grep_log_file(filepath, pattern)   — cross-platform case-insensitive grep
  _tail_lines(filepath, n=200)        — cross-platform tail
  _get_log_dirs()                     — candidate log-dir discovery
  _find_log_file(ds)                  — date-stamped log lookup

All four are heavily used across route modules (26 call sites in
routes/channels.py, routes/brain.py, routes/components.py, routes/health.py,
routes/infra.py, routes/overview.py). Re-exported from dashboard.py so
`_d._get_log_dirs()` etc. keep working without changes to routes/.

## Notes

- `_find_log_file` uses a late `import dashboard as _d` to reach the
  runtime-set `LOG_DIR` override (set from `--log-dir` / env). Same
  pattern routes/ already uses for cross-module state.
- dashboard.py had *two identical copies* of `_grep_log_file`,
  `_tail_lines`, and `_get_log_dirs` (at lines 137-180 and 6051-6094) —
  consequence of the embedded HTML-string-closing quirk that split the
  file's preamble. Consolidating into helpers/logs.py removes both
  copies and leaves a single canonical binding.
- Standalone `re as _re` and `tempfile as _tempfile` imports in
  dashboard.py are no longer strictly needed but were left in place —
  other code paths may reference them; follow-up PR can prune.

## E2E

- Unit tests on helpers/logs.py pass
- 39/39 API endpoints return 200 across every tab + all 13 route modules
- Zero tracebacks in server log
- `_tail_lines(gateway.log, 3)` returns 3 real lines end-to-end

## Sizes

  dashboard.py    : -80 lines (two duplicate sets of 3 funcs removed +
                               _find_log_file removed; 6-line re-export
                               import added; net -74)
  helpers/logs.py : new (82 lines with docstrings)
